+++ src/main/java/com/project/group4/models/json/IncludedNoteJson.java
@@ -1,4 +1,4 @@
-package com.project.group4.models.jsonframe;
+package com.project.group4.models.json;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
